test: split qbft timeout tests
